use relTol to compare double values.
