+++ mllib/src/test/scala/org/apache/spark/mllib/feature/NormalizerSuite.scala
@@ -23,6 +23,7 @@ import org.apache.spark.SparkFunSuite
 import org.apache.spark.mllib.linalg.{DenseVector, SparseVector, Vectors}
 import org.apache.spark.mllib.util.MLlibTestSparkContext
 import org.apache.spark.mllib.util.TestingUtils._
+import org.apache.spark.util.TestingUtils._
 
 class NormalizerSuite extends SparkFunSuite with MLlibTestSparkContext {
